split out array and stream
