+++ python/tests/test_nanoarrow.py
@@ -43,11 +43,13 @@ def test_c_schema_helper():
 
 
 def test_c_array_helper():
+    from nanoarrow.c_array import CArray
+
     array = na.allocate_c_array()
     assert na.c_array(array) is array
 
     array = na.c_array(pa.array([], pa.null()))
-    assert isinstance(array, na.c_lib.CArray)
+    assert isinstance(array, CArray)
 
     with pytest.raises(TypeError):
         na.c_array(1234)
@@ -62,9 +64,11 @@ def test_array_stream_helper():
 
 
 def test_array_view_helper():
+    from nanoarrow.c_array import CArrayView
+
     array = na.c_array(pa.array([1, 2, 3]))
     view = na.c_array_view(array)
-    assert isinstance(view, na.c_lib.CArrayView)
+    assert isinstance(view, CArrayView)
     assert na.c_array_view(view) is view
